Updated autoxref to do parts of exercises (13.3.b)
Now, autoxref works for references to parts of exercises, which are listed with lowercase Latin letters (13.3.b). It's a reasonable style question to ask if we should refer to parts of questions.

Best news yet: This change to `autoxref-treeprocess.rb` somehow fixed exercise reference labeling in PDFs, which were previously broken.