Updated information about the JCF and changed code style in dynamic data structure chapter
One change was to swap the `ArrayList` and `Vector` examples, since we want to emphasize that `ArrayList` is the typical tool.

Tried to make the code style close to what we want. Code style is going to be an annoying thing to fix since some code is directly in the book, not in a file that can be automatically formatted and because automatically formatting some code can change the size of the program, affecting places where code has been included in the book by line number.