Changes to PDF generation for consistency
Now, we're using the same body font and a very similar font for code when generating the PDF. Is it better? It matches the diagrams at least. Fonts are included in this commit.

We also updated the theming so that images and tables (and their captions) are always centered.

Unfortunately, updating tables meant many updates to many files.